+++ types/metasql.d.ts
@@ -1,0 +1,3 @@
+export class Database {
+  count(table: string, ...conditions: Array<object>): Promise<number>;
+}